pictureoptimizer2/requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
dj-database-url==0.5.0
dj-static==0.0.6
dj-stripe==1.2.3
Django==1.11.6
django-filter==2.0.0
django-toolbelt==0.0.1
djangorestframework==3.9.0
djangorestframework-filters==0.10.2
enum34==1.1.6
gunicorn==19.9.0
idna==2.5
ipaddress==1.0.22
jsonfield==2.0.2
psycopg2==2.7.6.1
psycopg2-binary==2.7.5
pycparser==2.19
pyOpenSSL==18.0.0
pytz==2018.9
requests==2.20.0
six==1.12.0
static3==0.7.0
stripe==2.17.0
urllib3==1.24.1
virtualenv==16.0.0